* master: (88 commits)
  changed
  use tags
  remove dead code
  fix names of server and project
  use first or create to tighten up the creation logic
  fix find_or_create_by in #630
  Some work on fixing #630
  Fixed config bug noted in concerto-debian #21
  Match the CanCan check with the delete call.
  Minor: Clean up screen auth setup page
  ++ isn't a thing.
  Bump to 0.6.0.HotelHummingbird.
  The logic for #620 wasnt working - this works, but maybe something more clever is warranted...
  Missed a few gems.
  Rebuild the frontend.
  Bump all gems.
  For #536, moved field configs into a container that is hidden by default if no field configs are set; changed JS so that modifying param data type also enables save button; removed confirmation on delete subscription button in subscriptions#manage because it is not applicable to the save model used on this page
  Missing </div> tag.
  Added zero-state banner when no active/visible feeds are available to close #620
  Fixed #623
  ...

Conflicts:
	Gemfile.lock
	db/schema.rb

+++ app/controllers/users_controller.rb
@@ -81,8 +81,13 @@ class UsersController < ApplicationController
     @user = User.find(params[:id])
     auth!
 
-    if !@user.destroy
-      flash[:notice] = t(:cannot_delete_last_admin)
+    if !@user.screens.empty?
+      # dont need to list screens because they'll see them on the user page
+      flash[:notice] = t(:user_owns_screens)
+    else
+      if !@user.destroy
+        flash[:notice] = t(:cannot_delete_last_admin)
+      end
     end
     respond_with(@user)
   end
